+++ packages/yourloops/test/integration/mock/dbl-communication.api.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021-2025, Diabeloop
+ * Copyright (c) 2025, Diabeloop
  *
  * All rights reserved.
  *
@@ -25,35 +25,22 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-export interface AppConfig {
-  VERSION: string
-  API_HOST: string
-  BRANDING: string
-  DOMAIN_NAME?: string
-  DEV: boolean
-  TEST: boolean
-  LATEST_TERMS?: string
-  LATEST_TRAINING?: string
-  LATEST_RELEASE?: string
-  SUPPORT_WEB_ADDRESS?: string
-  CONTACT_SUPPORT_WEB_URL: string
-  ASSETS_URL?: string | null
-  METRICS_SERVICE?: string | null
-  TERMS_PRIVACY_DATE?: string
-  STONLY_WID: string
-  COOKIE_BANNER_CLIENT_ID: string
-  SESSION_TIMEOUT: number
-  YLP820_BASAL_TIME: number
-  CBG_BUCKETS_ENABLED: boolean
-  ECPS_ENABLED: boolean
-  AUTH0_DOMAIN: string
-  AUTH0_ISSUER: string
-  AUTH0_CLIENT_ID: string
-  YLPZ_RA_LAD_FR: string
-  YLPZ_RA_LAD_EN: string
-  YLPZ_RA_LAD_NL: string
-  YLPZ_RA_LAD_IT: string
-  YLPZ_RA_LAD_ES: string
-  YLPZ_RA_LAD_DE: string
-  IDLE_TIMEOUT_MS: number
+import DblCommunicationApi from '../../../lib/dbl-communication/dbl-communication.api'
+import { BannerContent } from '../../../lib/dbl-communication/models/banner.model'
+import { InformationPage } from '../../../lib/dbl-communication/models/page.model'
+
+export const mockDblCommunicationApi = () => {
+  jest.spyOn(DblCommunicationApi, 'getInfoPage').mockResolvedValue(null)
+  jest.spyOn(DblCommunicationApi, 'getDblBanner').mockResolvedValue(null)
+}
+
+export const mockDblCommunicationApiBanner = (bannerContent: BannerContent) => {
+
+jest.spyOn(DblCommunicationApi, 'getInfoPage').mockResolvedValue(null)
+  jest.spyOn(DblCommunicationApi, 'getDblBanner').mockResolvedValue(bannerContent)
+}
+
+export const mockDblCommunicationApiPage = (infoPage: InformationPage) => {
+  jest.spyOn(DblCommunicationApi, 'getInfoPage').mockResolvedValue(infoPage)
+  jest.spyOn(DblCommunicationApi, 'getDblBanner').mockResolvedValue(null)
 }