+++ packages/yourloops/test/integration/mock/patient-login.mock.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022-2023, Diabeloop
+ * Copyright (c) 2022-2025, Diabeloop
  *
  * All rights reserved.
  *
@@ -35,10 +35,12 @@ import { UserRole } from '../../../lib/auth/models/enums/user-role.enum'
 import { mockUserApi } from './user.api.mock'
 import { mockPatientApiForPatients } from './patient.api.mock'
 import { Patient } from '../../../lib/patient/models/patient.model'
+import { mockDblCommunicationApi } from './dbl-communication.api'
 
 export const mockPatientLogin = (patient: Patient) => {
   mockAuth0Hook(UserRole.Patient, patient.userid)
   mockNotificationAPI()
+  mockDblCommunicationApi()
   mockDirectShareApi()
   mockTeamAPI()
   mockUserApi().mockUserDataFetch({ firstName: patient.profile.firstName, lastName: patient.profile.lastName })