+++ packages/yourloops/test/integration/pages/care-team-settings/care-team.patient.spec.tsx
@@ -39,6 +39,7 @@ import { patient1Id } from '../../data/patient.api.data'
 import { mockDataAPI } from '../../mock/data.api.mock'
 import { mockMedicalFilesAPI } from '../../mock/medical-files.api.mock'
 import { mockChatAPI } from '../../mock/chat.api.mock'
+import { mockDblCommunicationApi } from '../../mock/dbl-communication.api'
 
 describe('Patient care team settings page', () => {
   const firstName = 'Jacques'
@@ -55,6 +56,7 @@ describe('Patient care team settings page', () => {
     mockDataAPI()
     mockMedicalFilesAPI(mySecondTeamId, mySecondTeamName)
     mockChatAPI()
+    mockDblCommunicationApi()
   })
 
   const renderCareTeamSettingsPage = async (route: string) => {